Moved header component and renamed with capital first letter.
