+++ Client/src/App.tsx
@@ -1,21 +1,23 @@
 import { Box } from "@mui/material";
+import { ThemeProvider } from "@mui/material/styles";
 import { Outlet } from "react-router-dom";
 import Footer from "./components/Footer";
-import Header from "./header";
+import Header from "./components/Header";
 import { theme } from "./theme";
-import { ThemeProvider } from "@mui/material/styles";
 
 export default function App() {
   return (
     <ThemeProvider theme={theme}>
-      <Box sx={{height: '100%'}}>
+      <Box sx={{ height: "100%" }}>
         <header>
           <Header />
         </header>
         <main>
           <Outlet />
         </main>
-        <footer><Footer /></footer>
+        <footer>
+          <Footer />
+        </footer>
       </Box>
     </ThemeProvider>
   );
